add components
There are three components in this app, a house, an error,
and a loading component.

A house component renders a single house.

A loading component renders a message that informs
users that React is fetching a json file

an error component renders a message that informs
users that the application fails to fetch the json file